# Conflicts:
#	src/main/java/com/amr/project/service/abstracts/ItemService.java
#	src/main/resources/import.sql

+++ src/main/java/com/amr/project/webapp/rest_controller/CartRestController.java
@@ -50,11 +50,12 @@ public class CartRestController {
 
     @Transactional
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<CartItemDto>> getAllCartItemsByUser(Principal principal) {
-        if(principal == null || principal instanceof AnonymousAuthenticationToken) {
+    public ResponseEntity<List<CartItemDto>> getAllCartItemsByUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!authentication.isAuthenticated() || (authentication instanceof AnonymousAuthenticationToken)) {
             throw new AccessDeniedException("Вам нужно авторизоваться для доступа к корзине");
         }
-        User user = userService.findByUsername(principal.getName()).get();
+        User user = userService.findByUsername(authentication.getName()).get();
         List<CartItem> cartItems = cartItemService.findByUser(user);
         List<CartItemDto> cartItemsDto = cartItems.stream().map(c -> cartItemMapper.cartItemToDto(c)).collect(Collectors.toList());
         LOGGER.info(String.format("Пользователь с id %d успешно получил список товаров в корзине", user.getId()));
@@ -64,13 +65,25 @@ public class CartRestController {
     @Transactional
     @PatchMapping("/update/{id}")
     public ResponseEntity<Void> updateCartItemQuantity(@PathVariable("id") Long id, @RequestBody CartItemDto cartItem) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!authentication.isAuthenticated() || (authentication instanceof AnonymousAuthenticationToken)) {
+            throw new AccessDeniedException("Вам нужно авторизоваться для доступа к корзине");
+        }
+        if (cartItem.getId() != id ) {
+            throw new IllegalArgumentException(String.format("Переданный параметр строки id %d не соответствует" +
+                    "параметру json id %d", id, cartItem.getId()));
+        }
         cartItemService.getByKey(id).setQuantity(cartItem.getQuantity());
         return ResponseEntity.ok().body(null);
     }
 
     @Transactional
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> updateCartItemQuantity(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCartItem(@PathVariable Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!authentication.isAuthenticated() || (authentication instanceof AnonymousAuthenticationToken)) {
+            throw new AccessDeniedException("Вам нужно авторизоваться для доступа к корзине");
+        }
         cartItemService.deleteByKeyCascadeIgnore(id);
         return ResponseEntity.ok().body(null);
     }
@@ -79,7 +92,6 @@ public class CartRestController {
     @PostMapping(value = "/add")
     public ResponseEntity<Void> addItemToCart(@RequestBody CartItemDto cartItemDto) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        System.out.println(authentication.getName());
         if(authentication == null || authentication instanceof AnonymousAuthenticationToken) {
             throw new AccessDeniedException("Вам нужно авторизоваться для доступа к корзине");
         }
@@ -91,6 +103,7 @@ public class CartRestController {
                 cartItemDto.getShop().getId()).isPresent()) {
             cartItem = cartItemService.findByItemAndShopAndUser(cartItemDto.getItem().getId(), cartItemDto.getUser().getId(),
                     cartItemDto.getShop().getId()).get();
+            cartItemDto.setQuantity(cartItem.getQuantity() + cartItemDto.getQuantity());
             updateCartItemQuantity(cartItem.getId(), cartItemDto);
         } else {
             cartItem = cartItemMapper.dtoToCartItem(cartItemDto);
